refactor(render): D1 thin-shell — move 7 leaf command groups to ts-render-services (#91)
Move business_logic + types for compiler_settings_commands, service_container,
service, gpu, cache, preview_advanced, info from ts-render-tauri into the
Tauri-free ts-render-services crate.

ts-render-tauri now holds only #[tauri::command] thin wrappers that delegate to
ts_render_services::<group>::commands_impl or re-exported business_logic fns.
All public command signatures are unchanged — app_builder.rs/specta_export.rs
require no edits.

Adds sysinfo, num_cpus, os_info, base64 to ts-render-services deps for the
info and preview_advanced groups.

Checks: cargo check -p ts-render-services ✓ | -p ts-render-tauri ✓
Tests:  210 passed (ts-render-services) | 493 passed (ts-render-tauri)
