+++ crates/ts-render-services/Cargo.toml
@@ -24,6 +24,13 @@ uuid = { version = "1", features = ["serde", "v4"] }
 once_cell = "1"
 chrono = { version = "0.4", features = ["serde"] }
 log = "0.4"
+# compiler_settings_commands-группа (D1): рекомендованное число параллельных задач.
+num_cpus = "1.17"
+# preview_advanced-группа (D1): base64-декод превью-кадров.
+base64 = "0.22"
+# info-группа (D1): сбор системной информации (диски/CPU/ОС).
+sysinfo = "0.37.2"
+os_info = "3.12.0"
 
 [dev-dependencies]
 tempfile = "3"
