+++ crates/ts-render-services/src/cache/mod.rs
@@ -1,0 +1,7 @@
+//! Модуль cache — управление кэшем рендеринга
+
+pub mod business_logic;
+pub mod commands_impl;
+pub mod types;
+
+pub use types::*;
