+++ crates/ts-render-services/src/compiler_settings_commands/mod.rs
@@ -1,0 +1,6 @@
+//! Модуль для настроек компилятора
+
+pub mod business_logic;
+pub mod types;
+
+pub use types::*;
