+++ crates/ts-render-services/src/gpu/mod.rs
@@ -1,0 +1,7 @@
+//! Модуль gpu — работа с GPU для рендеринга
+
+pub mod business_logic;
+pub mod commands_impl;
+pub mod types;
+
+pub use types::*;
