+++ crates/ts-render-services/src/info/mod.rs
@@ -1,0 +1,7 @@
+//! Модуль info — информация о системе и компиляторе
+
+pub mod business_logic;
+pub mod commands_impl;
+pub mod types;
+
+pub use types::*;
