+++ crates/ts-render-services/src/preview_advanced/mod.rs
@@ -1,0 +1,7 @@
+//! Модуль preview_advanced — расширенные команды превью
+
+pub mod business_logic;
+pub mod commands_impl;
+pub mod types;
+
+pub use types::*;
