+++ crates/ts-render-services/src/service/mod.rs
@@ -1,0 +1,6 @@
+//! Модуль service — управление сервисами рендеринга
+
+pub mod business_logic;
+pub mod types;
+
+pub use types::*;
