+++ crates/ts-render-services/src/service_container/mod.rs
@@ -1,0 +1,6 @@
+//! Модуль для работы с ServiceContainer и метриками
+
+pub mod business_logic;
+pub mod types;
+
+pub use types::*;
