+++ crates/ts-render-tauri/src/compiler_settings_commands/commands.rs
@@ -1,17 +1,17 @@
 //! Thin команды для настроек компилятора
 
-use super::business_logic;
 use super::types::*;
 use tauri::State;
 use ts_render::video_compiler::core::error::Result;
 use ts_render_services::VideoCompilerState;
+use ts_render_services::compiler_settings_commands::business_logic as bl;
 
 /// Получить настройки компилятора
 #[tauri::command]
 pub async fn get_compiler_settings_advanced(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<CompilerSettings> {
-  Ok(business_logic::create_default_settings())
+  Ok(bl::create_default_settings())
 }
 
 /// Обновить настройки компилятора
@@ -20,7 +20,7 @@ pub async fn update_compiler_settings_advanced(
   settings: CompilerSettings,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<bool> {
-  business_logic::validate_compiler_settings(&settings)?;
+  bl::validate_compiler_settings(&settings)?;
   log::info!("Updating compiler settings: {settings:?}");
   Ok(true)
 }
@@ -31,7 +31,7 @@ pub async fn set_ffmpeg_path_advanced(
   path: String,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<bool> {
-  business_logic::validate_ffmpeg_path(&path)?;
+  bl::validate_ffmpeg_path(&path)?;
   log::info!("Setting FFmpeg path to: {path}");
   Ok(true)
 }
@@ -42,7 +42,7 @@ pub async fn set_parallel_jobs_advanced(
   jobs: usize,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<bool> {
-  business_logic::validate_parallel_jobs(jobs)?;
+  bl::validate_parallel_jobs(jobs)?;
   log::info!("Setting parallel jobs to: {jobs}");
   Ok(true)
 }
@@ -53,7 +53,7 @@ pub async fn set_memory_limit_advanced(
   limit_mb: usize,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<bool> {
-  business_logic::validate_memory_limit(limit_mb)?;
+  bl::validate_memory_limit(limit_mb)?;
   log::info!("Setting memory limit to: {limit_mb} MB");
   Ok(true)
 }
@@ -64,7 +64,7 @@ pub async fn set_temp_directory_advanced(
   directory: String,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<bool> {
-  business_logic::validate_temp_directory(&directory)?;
+  bl::validate_temp_directory(&directory)?;
   log::info!("Setting temp directory to: {directory}");
   Ok(true)
 }
@@ -75,7 +75,7 @@ pub async fn set_log_level_advanced(
   level: String,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<bool> {
-  business_logic::validate_log_level(&level)?;
+  bl::validate_log_level(&level)?;
   log::info!("Setting log level to: {level}");
   Ok(true)
 }
@@ -86,7 +86,7 @@ pub async fn reset_compiler_settings_advanced(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<CompilerSettings> {
   log::info!("Resetting compiler settings to defaults");
-  Ok(business_logic::create_default_settings())
+  Ok(bl::create_default_settings())
 }
 
 /// Получить рекомендуемые настройки
@@ -94,14 +94,14 @@ pub async fn reset_compiler_settings_advanced(
 pub async fn get_recommended_settings_advanced(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<RecommendedSettings> {
-  Ok(business_logic::create_recommended_settings())
+  Ok(bl::create_recommended_settings())
 }
 
 /// Экспортировать настройки
 #[tauri::command]
 pub async fn export_settings_advanced(_state: State<'_, VideoCompilerState>) -> Result<String> {
-  let settings = business_logic::create_default_settings();
-  business_logic::export_settings_to_json(&settings)
+  let settings = bl::create_default_settings();
+  bl::export_settings_to_json(&settings)
 }
 
 /// Импортировать настройки
@@ -110,7 +110,7 @@ pub async fn import_settings_advanced(
   settings_json: String,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<CompilerSettings> {
-  business_logic::import_settings_from_json(&settings_json)
+  bl::import_settings_from_json(&settings_json)
 }
 
 /// Получить пресеты качества
@@ -118,5 +118,5 @@ pub async fn import_settings_advanced(
 pub async fn get_quality_presets_advanced(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<Vec<QualityPreset>> {
-  Ok(business_logic::create_quality_presets())
+  Ok(bl::create_quality_presets())
 }
