+++ crates/ts-render-tauri/src/gpu/commands.rs
@@ -1,21 +1,15 @@
 //! Tauri команды для работы с GPU
 
-use super::business_logic;
 use super::types::*;
-use ts_render::video_compiler::core::error::{Result, VideoCompilerError};
-use ts_render::video_compiler::core::gpu::GpuDetector;
+use ts_render::video_compiler::core::error::Result;
 use ts_render_services::VideoCompilerState;
 use tauri::State;
+use ts_render_services::gpu::commands_impl as impl_;
 
 /// Обнаружить доступные GPU
 #[tauri::command]
 pub async fn detect_gpus(state: State<'_, VideoCompilerState>) -> Result<Vec<GpuInfo>> {
-  let gpu_service = state
-    .services
-    .get_gpu_service()
-    .ok_or_else(|| VideoCompilerError::validation("GpuService не найден"))?;
-
-  gpu_service.detect_gpus().await
+  impl_::detect_gpus(&state).await
 }
 
 /// Получить возможности GPU
@@ -24,12 +18,7 @@ pub async fn get_gpu_capabilities(
   _gpu_index: usize,
   state: State<'_, VideoCompilerState>,
 ) -> Result<ts_render_services::services::gpu_service::GpuCapabilities> {
-  let gpu_service = state
-    .services
-    .get_gpu_service()
-    .ok_or_else(|| VideoCompilerError::validation("GpuService не найден"))?;
-
-  gpu_service.get_capabilities().await
+  impl_::get_gpu_capabilities(&state).await
 }
 
 /// Проверить поддержку аппаратного ускорения
@@ -37,31 +26,19 @@ pub async fn get_gpu_capabilities(
 pub async fn check_hardware_acceleration_support(
   state: State<'_, VideoCompilerState>,
 ) -> Result<bool> {
-  let gpu_service = state
-    .services
-    .get_gpu_service()
-    .ok_or_else(|| VideoCompilerError::validation("GpuService не найден"))?;
-  gpu_service.check_hardware_acceleration().await
+  impl_::check_hardware_acceleration_support(&state).await
 }
 
 /// Проверить доступность GPU (алиас для check_hardware_acceleration_support)
 #[tauri::command]
 pub async fn check_gpu_availability(state: State<'_, VideoCompilerState>) -> Result<bool> {
-  check_hardware_acceleration_support(state).await
+  impl_::check_hardware_acceleration_support(&state).await
 }
 
 /// Получить рекомендуемый GPU для рендеринга
 #[tauri::command]
 pub async fn get_recommended_gpu(state: State<'_, VideoCompilerState>) -> Result<Option<GpuInfo>> {
-  let gpu_service = state
-    .services
-    .get_gpu_service()
-    .ok_or_else(|| VideoCompilerError::validation("GpuService не найден"))?;
-
-  // Получаем рекомендуемый кодировщик и конвертируем в GpuInfo
-  let recommended_encoder = gpu_service.get_recommended_encoder().await?;
-
-  Ok(recommended_encoder.map(business_logic::create_recommended_gpu_info))
+  impl_::get_recommended_gpu(&state).await
 }
 
 /// Установить предпочитаемый GPU для рендеринга
@@ -70,11 +47,7 @@ pub async fn set_preferred_gpu(
   _gpu_index: usize,
   state: State<'_, VideoCompilerState>,
 ) -> Result<()> {
-  // Просто включаем аппаратное ускорение
-  let mut settings = state.settings.write().await;
-  settings.hardware_acceleration = true;
-
-  Ok(())
+  impl_::set_preferred_gpu(&state).await
 }
 
 /// Включить/выключить аппаратное ускорение
@@ -83,9 +56,7 @@ pub async fn set_hardware_acceleration(
   enabled: bool,
   state: State<'_, VideoCompilerState>,
 ) -> Result<()> {
-  let mut settings = state.settings.write().await;
-  settings.hardware_acceleration = enabled;
-  Ok(())
+  impl_::set_hardware_acceleration(&state, enabled).await
 }
 
 /// Получить текущий статус использования GPU
@@ -93,12 +64,7 @@ pub async fn set_hardware_acceleration(
 pub async fn get_gpu_usage_status(
   state: State<'_, VideoCompilerState>,
 ) -> Result<serde_json::Value> {
-  let settings = state.settings.read().await;
-
-  let status =
-    business_logic::create_gpu_usage_status(settings.hardware_acceleration, None, None, 0);
-
-  Ok(serde_json::to_value(status)?)
+  impl_::get_gpu_usage_status(&state).await
 }
 
 /// Протестировать производительность GPU
@@ -107,8 +73,8 @@ pub async fn benchmark_gpu(
   _gpu_index: usize,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<serde_json::Value> {
-  // Бенчмарк GPU не реализован, возвращаем заглушку
-  let result = business_logic::create_benchmark_result_stub();
+  use ts_render_services::gpu::business_logic::create_benchmark_result_stub;
+  let result = create_benchmark_result_stub();
   Ok(serde_json::to_value(result)?)
 }
 
@@ -118,33 +84,22 @@ pub async fn get_gpu_supported_codecs(
   _gpu_index: usize,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<Vec<String>> {
-  // Возвращаем стандартные кодеки для GPU
-  Ok(business_logic::get_standard_gpu_codecs())
+  use ts_render_services::gpu::business_logic::get_standard_gpu_codecs;
+  Ok(get_standard_gpu_codecs())
 }
 
 /// Автоматически выбрать лучший GPU
 #[tauri::command]
 pub async fn auto_select_gpu(state: State<'_, VideoCompilerState>) -> Result<Option<usize>> {
-  let gpu_service = state
-    .services
-    .get_gpu_service()
-    .ok_or_else(|| VideoCompilerError::validation("GpuService не найден"))?;
-  let recommended = gpu_service.get_recommended_encoder().await?;
-
-  if recommended.is_some() {
-    let mut settings = state.settings.write().await;
-    settings.hardware_acceleration = true;
-    Ok(business_logic::determine_gpu_index(true))
-  } else {
-    Ok(None)
-  }
+  impl_::auto_select_gpu(&state).await
 }
 
 /// Получить детальную информацию о GPU кодировщике
 #[tauri::command]
 pub async fn get_gpu_encoder_details(encoder_type: String) -> Result<serde_json::Value> {
-  let encoder = business_logic::parse_encoder_type(&encoder_type);
-  let details = business_logic::create_gpu_encoder_details(&encoder, &encoder_type);
+  use ts_render_services::gpu::business_logic::{create_gpu_encoder_details, parse_encoder_type};
+  let encoder = parse_encoder_type(&encoder_type);
+  let details = create_gpu_encoder_details(&encoder, &encoder_type);
   Ok(serde_json::to_value(details)?)
 }
 
@@ -153,19 +108,5 @@ pub async fn get_gpu_encoder_details(encoder_type: String) -> Result<serde_json:
 pub async fn get_gpu_capabilities_full(
   state: State<'_, VideoCompilerState>,
 ) -> Result<serde_json::Value> {
-  let detector = GpuDetector::new(state.ffmpeg_path.read().await.clone());
-  let capabilities = detector.get_gpu_capabilities().await?;
-
-  let info = business_logic::create_gpu_capabilities_info(
-    capabilities
-      .available_encoders
-      .iter()
-      .map(|e| format!("{:?}", e))
-      .collect(),
-    capabilities.hardware_acceleration_supported,
-    capabilities.recommended_encoder.map(|e| format!("{:?}", e)),
-    capabilities.current_gpu.map(|g| g.name),
-  );
-
-  Ok(serde_json::to_value(info)?)
+  impl_::get_gpu_capabilities_full(&state).await
 }
