+++ crates/ts-render-tauri/src/service/commands.rs
@@ -1,14 +1,15 @@
 //! Tauri команды для управления сервисами
 
-use super::{business_logic, types::*};
+use super::types::*;
 use tauri::State;
 use ts_render::video_compiler::error::Result;
 use ts_render_services::VideoCompilerState;
+use ts_render_services::service::business_logic as bl;
 
 /// Получить активные задачи
 #[tauri::command]
 pub async fn get_active_jobs(state: State<'_, VideoCompilerState>) -> Result<Vec<String>> {
-  Ok(business_logic::get_active_job_ids(&state).await)
+  Ok(bl::get_active_job_ids(&state).await)
 }
 
 /// Получить прогресс рендеринга
@@ -17,7 +18,7 @@ pub async fn get_render_progress(
   job_id: String,
   state: State<'_, VideoCompilerState>,
 ) -> Result<f64> {
-  business_logic::get_job_render_progress(&state, &job_id).await
+  bl::get_job_render_progress(&state, &job_id).await
 }
 
 /// Получить статистику рендеринга
@@ -26,7 +27,7 @@ pub async fn get_render_statistics(
   job_id: String,
   state: State<'_, VideoCompilerState>,
 ) -> Result<RenderStatistics> {
-  business_logic::create_render_statistics(&state, &job_id).await
+  bl::create_render_statistics(&state, &job_id).await
 }
 
 /// Получить информацию об источниках входных данных
@@ -34,7 +35,7 @@ pub async fn get_render_statistics(
 pub async fn get_input_sources_info(
   project_schema: ts_render::video_compiler::schema::ProjectSchema,
 ) -> Result<InputSourcesResult> {
-  Ok(business_logic::analyze_input_sources(&project_schema))
+  Ok(bl::analyze_input_sources(&project_schema))
 }
 
 /// Обновить время доступа к проекту
@@ -53,7 +54,7 @@ pub async fn set_preview_ffmpeg_path(
   state: State<'_, VideoCompilerState>,
 ) -> Result<()> {
   // Проверяем путь
-  business_logic::validate_ffmpeg_path(&path)?;
+  bl::validate_ffmpeg_path(&path)?;
 
   // Обновляем путь
   let mut ffmpeg_path = state.ffmpeg_path.write().await;
@@ -68,7 +69,7 @@ pub async fn get_all_service_metrics(
   state: State<'_, VideoCompilerState>,
 ) -> Result<AllServiceMetrics> {
   log::debug!("Getting all service metrics");
-  Ok(business_logic::create_all_service_metrics(&state).await)
+  Ok(bl::create_all_service_metrics(&state).await)
 }
 
 /// Получить метрики конкретного сервиса
@@ -78,7 +79,7 @@ pub async fn get_specific_service_metrics(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<SpecificServiceMetrics> {
   log::debug!("Getting metrics for service: {service_name}");
-  business_logic::create_specific_service_metrics(&service_name)
+  bl::create_specific_service_metrics(&service_name)
 }
 
 /// Очистить завершенные задачи
@@ -92,7 +93,7 @@ pub async fn cleanup_completed_jobs(
     older_than_hours.unwrap_or(24)
   );
 
-  let result = business_logic::cleanup_completed_jobs_logic(&state, older_than_hours).await;
+  let result = bl::cleanup_completed_jobs_logic(&state, older_than_hours).await;
   Ok(result.removed_count)
 }
 
@@ -102,7 +103,7 @@ pub async fn get_services_health(
   state: State<'_, VideoCompilerState>,
 ) -> Result<ServicesHealthStatus> {
   log::debug!("Checking services health");
-  Ok(business_logic::create_services_health_status(&state).await)
+  Ok(bl::create_services_health_status(&state).await)
 }
 
 /// Перезапустить сервис
@@ -111,6 +112,6 @@ pub async fn restart_service(
   service_name: String,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<()> {
-  let service_type = business_logic::validate_service_for_restart(&service_name)?;
-  business_logic::restart_service_logic(service_type)
+  let service_type = bl::validate_service_for_restart(&service_name)?;
+  bl::restart_service_logic(service_type)
 }
