+++ crates/ts-render-tauri/src/service_container/commands.rs
@@ -1,6 +1,6 @@
 //! Tauri команды для работы с ServiceContainer и MetricsRegistry
 
-use super::{business_logic, types::*};
+use super::types::*;
 use tauri::State;
 use ts_render::video_compiler::core::error::Result;
 use ts_render_services::VideoCompilerState;
@@ -10,7 +10,7 @@ use ts_render_services::VideoCompilerState;
 pub async fn get_project_service_info_command(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<ProjectServiceInfo> {
-  business_logic::get_project_service_info_logic().await
+  ts_render_services::service_container::business_logic::get_project_service_info_logic().await
 }
 
 /// Получить метрики конкретного сервиса
@@ -19,7 +19,8 @@ pub async fn get_service_metrics_detailed(
   params: GetServiceMetricsParams,
   _state: State<'_, VideoCompilerState>,
 ) -> Result<ServiceMetricsInfo> {
-  business_logic::get_service_metrics_detailed_logic(&params).await
+  ts_render_services::service_container::business_logic::get_service_metrics_detailed_logic(&params)
+    .await
 }
 
 /// Получить все сводки метрик
@@ -27,7 +28,7 @@ pub async fn get_service_metrics_detailed(
 pub async fn get_all_metrics_summaries_command(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<AllMetricsSummary> {
-  business_logic::get_all_metrics_summaries_logic().await
+  ts_render_services::service_container::business_logic::get_all_metrics_summaries_logic().await
 }
 
 /// Экспортировать метрики в формате Prometheus
@@ -35,5 +36,5 @@ pub async fn get_all_metrics_summaries_command(
 pub async fn export_prometheus_detailed(
   _state: State<'_, VideoCompilerState>,
 ) -> Result<PrometheusExport> {
-  business_logic::export_prometheus_detailed_logic().await
+  ts_render_services::service_container::business_logic::export_prometheus_detailed_logic().await
 }
